// app/src/types/type.ts
export interface ProductFilterFormValues {
  category?: string;
  minPrice?: string;
  maxPrice?: string;
  inStock?: boolean;
  sortBy?: 'name' | 'price' | 'createdAt';
  sortOrder?: 'asc' | 'desc';
}
export interface ProductFilterApiParams {
  page?: number;
  limit?: number;
  category?: string;
  name?: string; // Adicionado para busca por nome
  minPrice?: number;
  maxPrice?: number;
  inStock?: boolean;
  sortBy?: 'name' | 'price' | 'createdAt' | 'salePrice';
  sortOrder?: 'asc' | 'desc';
  collection?: string;
  hasDiscount?: boolean;
  searchTerm?: string; // Termo geral de busca
}